// frontend/src/app/models/publisher.model.ts
export class Publisher {
    id?: any;
    name?: String;
    published?: String
    showCount?: any;
    apiKey?: String;
    podApiKey?: String;
    email?: String;
    website?: String
    instance?: String;
    type?: String;
    user?: String;
    agency?: any;
    awsApikey?: String;
    awsApiUsagePlan?: String;
    imported?: Boolean;
    currentUsage?: number;
    remainingUsage?: number;

}